Refactor OrderItemDAO to temporarily remove subtotal handling in update and create methods; update Order model to replace voucherCode and discountAmount with voucherId; remove unused fields from OrderItem model; update OrderStatus enum to reflect new statuses and add Vietnamese display names; delete obsolete controllers for Order, Product, Review, and Voucher; implement new RootServlet and AdminController for improved API structure; add Auth, Cart, Order, Product, Review, and Voucher controllers with detailed endpoint documentation.
