+++ backend/src/main/java/com/utephonehub/controller/RootServlet.java
@@ -1,4 +1,4 @@
-package com.utephonehub;
+package com.utephonehub.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
