+++ backend/src/main/java/com/utephonehub/model/order/OrderItem.java
@@ -1,7 +1,6 @@
 package com.utephonehub.model.order;
 
 import java.math.BigDecimal;
-import java.sql.Timestamp;
 
 /**
  * Entity class representing an OrderItem in the database
@@ -12,8 +11,6 @@ public class OrderItem {
     private int productId;
     private int quantity;
     private BigDecimal price; // Price at the time of order (not current price)
-    private BigDecimal subtotal; // quantity * price
-    private Timestamp createdAt;
 
     // Default constructor
     public OrderItem() {}
@@ -24,8 +21,6 @@ public class OrderItem {
         this.productId = productId;
         this.quantity = quantity;
         this.price = price;
-        this.subtotal = price.multiply(BigDecimal.valueOf(quantity));
-        this.createdAt = new Timestamp(System.currentTimeMillis());
     }
 
     // Getters and Setters
@@ -51,10 +46,6 @@ public class OrderItem {
 
     public void setQuantity(int quantity) {
         this.quantity = quantity;
-        // Recalculate subtotal when quantity changes
-        if (this.price != null) {
-            this.subtotal = this.price.multiply(BigDecimal.valueOf(quantity));
-        }
     }
 
     public BigDecimal getPrice() {
@@ -63,35 +54,6 @@ public class OrderItem {
 
     public void setPrice(BigDecimal price) {
         this.price = price;
-        // Recalculate subtotal when price changes
-        if (price != null) {
-            this.subtotal = price.multiply(BigDecimal.valueOf(this.quantity));
-        }
-    }
-
-    public BigDecimal getSubtotal() {
-        return subtotal;
-    }
-
-    public void setSubtotal(BigDecimal subtotal) {
-        this.subtotal = subtotal;
-    }
-
-    public Timestamp getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Timestamp createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    /**
-     * Calculate subtotal based on current quantity and price
-     */
-    public void calculateSubtotal() {
-        if (this.price != null) {
-            this.subtotal = this.price.multiply(BigDecimal.valueOf(this.quantity));
-        }
     }
 
     @Override
@@ -101,8 +63,6 @@ public class OrderItem {
                 ", productId=" + productId +
                 ", quantity=" + quantity +
                 ", price=" + price +
-                ", subtotal=" + subtotal +
-                ", createdAt=" + createdAt +
                 '}';
     }
 }
