+++ backend/src/main/java/com/utephonehub/model/order/OrderStatus.java
@@ -4,13 +4,11 @@ package com.utephonehub.model.order;
  * Enum representing different states of an Order
  */
 public enum OrderStatus {
-    PENDING("Chờ xử lý"),
-    CONFIRMED("Đã xác nhận"), 
-    PREPARING("Đang chuẩn bị"),
-    SHIPPING("Đang giao hàng"),
-    DELIVERED("Đã giao hàng"),
-    CANCELLED("Đã hủy"),
-    RETURNED("Đã trả hàng");
+    PENDING("pending"),
+    PROCESSING("processing"), 
+    SHIPPED("shipped"),
+    COMPLETED("completed"),
+    CANCELLED("cancelled");
 
     private final String displayName;
 
@@ -34,6 +32,20 @@ public enum OrderStatus {
             return OrderStatus.valueOf(status.toUpperCase());
         } catch (IllegalArgumentException e) {
             return PENDING;
+        }
+    }
+
+    /**
+     * Get Vietnamese display name for status
+     */
+    public String getVietnameseDisplayName() {
+        switch (this) {
+            case PENDING: return "Chờ xử lý";
+            case PROCESSING: return "Đang xử lý";
+            case SHIPPED: return "Đang giao hàng";
+            case COMPLETED: return "Hoàn thành";
+            case CANCELLED: return "Đã hủy";
+            default: return displayName;
         }
     }
 }
